[FIX] Don't show follow options if the user isn't present

+++ app/helpers/motions_helper.rb
@@ -71,7 +71,7 @@ module MotionsHelper
     if mo_po.destroy?
       link_items << link_item(t('destroy'), motion_path(motion, destroy: true), data: {confirm: t('destroy_confirmation'), method: 'delete', 'skip-pjax' => 'true'}, fa: 'close')
     end
-    if active_for_user?(:notifications, current_user)
+    if current_profile && active_for_user?(:notifications, current_user)
       if current_profile.following?(motion)
         link_items << link_item(t('forums.unfollow'), follows_path(motion_id: motion.id), fa: 'bell-slash', data: {method: 'delete', 'skip-pjax' => 'true'})
       else
